# app/docs/openapi.json
{
  "openapi" : "3.1.0",
  "paths" : {
...
    "/internal/accounts/{userId}/nodes/{nodeId}" : {
      "get" : {
        "parameters" : [ {
          "name" : "nodeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InternalNodeDto"
                }
              }
            }
          }
        },
        "summary" : "Get Node",
        "tags" : [ "Internal Node Resource" ]
      }
    },
...
    "/internal/folders" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateFolderRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InternalNodeIdDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Create Folder",
        "tags" : [ "Internal Node Resource" ]
      }
    },
    "/internal/links" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreatePublicLinkRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublicLinkDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Create Public Link",
        "tags" : [ "Internal Node Resource" ]
      }
    },
    "/internal/nodes" : {
      "delete" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeleteAllRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeleteAllResponse"
                }
              }
            }
          }
        },
        "summary" : "Delete All Nodes And Blobs",
        "tags" : [ "Internal Node Resource" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "CreateFolderRequest" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "string"
          },
          "destinationId" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
      "CreatePublicLinkRequest" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "string"
          },
          "nodeId" : {
            "type" : "string"
          }
        }
      },
      "DeleteAllRequest" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "string"
          }
        }
      },
      "DeleteAllResponse" : {
        "type" : "object",
        "properties" : {
          "deleted" : {
            "type" : "boolean"
          }
        }
      },
      "InternalNodeDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "extension" : {
            "type" : "string"
          },
          "mimeType" : {
            "type" : "string"
          },
          "size" : {
            "type" : "integer",
            "format" : "int64"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          },
          "updatedAt" : {
            "type" : "integer",
            "format" : "int64"
          },
          "owner" : {
            "$ref" : "#/components/schemas/OwnerDto"
          },
          "parent" : {
            "$ref" : "#/components/schemas/ParentDto"
          },
          "permissions" : {
            "$ref" : "#/components/schemas/PermissionsDto"
          }
        }
      },
      "InternalNodeIdDto" : {
        "type" : "object",
        "properties" : {
          "nodeId" : {
            "type" : "string"
          }
        }
      },
      "OwnerDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          }
        }
      },
      "ParentDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          }
        }
      },
      "PermissionsDto" : {
        "type" : "object",
        "properties" : {
          "canWriteFile" : {
            "type" : "boolean"
          }
        }
      },
      "PublicLinkDto" : {
        "type" : "object",
        "properties" : {
          "url" : {
            "type" : "string"
          }
        }
      }
    }
  },
  "info" : {
    "title" : "carbonio-files-ce-app API",
    "version" : "unversioned"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}